moved mp test file somehow extremly slow
