+++ SenLAST/test_env/mp_test_file.py
@@ -1,0 +1,98 @@
+from pathos import multiprocessing as mp
+from datetime import datetime
+import os
+import rasterio as rio
+import rasterio.mask
+import numpy as np
+import fiona
+import shutil
+
+start_time = datetime.now()
+
+### ----- TIFF Data ----- ###
+## Jonas Folder:
+# directory = "F:/Sentinel_3/S3_Daten-20200622T143538Z-001/S3_Daten"
+## Marlin Folder:
+directory = "F:/GEO411_data/Sentinel_Daten/"
+
+### ----- Shapefile Data ----- ###
+## Shapefile Marlin:
+shapefolder = "F:/GEO411_data/Daten_Sandra/new/"
+shapefile = "Stationen_Thüringen_Umland_3x3box.shp"
+## Shapefile Jonas:
+# shapefolder = "C:/Users/jz199/Documents/Studium/Master/2. Semester/Vorlesungsmitschriften/GEO411 - Landschaftsmanagement und Fernerkundung/Auszug_Daten_SandraBauer_MA/Auszug_Daten_SandraBauer_MA/"
+#shapefile = "Stationen_Thüringen_Umland_3x3box.shp"
+
+shape_path = shapefolder + shapefile
+
+
+def extract_files_to_list(path_to_folder):
+    #print("a")
+    """
+    finds all .tif-files in the corresponding directory
+    :return:
+    """
+    new_list = []
+
+    for filename in os.listdir(path_to_folder):
+        if filename.endswith(".tif"):
+            new_list.append(os.path.join(path_to_folder, filename))
+        else:
+            continue
+    return new_list
+
+
+### this function needs to be generalized ###
+def import_polygons(shape_path):
+    #print("b")
+    """
+    imports the 3x3km polygons of the DWD weather stations
+    :return:
+    """
+    shape_list = []
+    acitve_shapefile = fiona.open(shape_path, "r")
+    for i in range(0,len(list(acitve_shapefile))-1):
+        shapes = [feature["geometry"] for feature in acitve_shapefile]
+        shape_list.append(shapes)
+    #print(shape_list)
+    return shape_list
+
+
+def eliminate_nanoverlap(file_list):
+    #print("c")
+    """
+    eliminates the scenes which would'nt match with all weather stations
+    :return:
+    """
+    #import_list = import_polygons(shape_path=shape_path)
+    import_list = [[{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}]]
+    # file_list = extract_files_to_list(path_to_folder=directory)
+    tifs_selected = directory + "/selected"
+    #files = ['F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-03_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-03_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-04_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-05_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-05_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-07_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-07_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-08_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-08_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-11_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-11_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-12_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-12_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-15_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-15_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-16_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-16_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-18_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-19_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-19_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-20_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-20_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-22_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-22_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-23_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-23_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-24_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-24_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-26_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-26_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-27_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-27_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-28_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-28_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-04_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-22_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-22_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-23_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-23_20h50m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-24_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-24_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-26_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-26_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-27_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-27_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-28_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-30_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-30_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-31_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-08-31_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-03_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-03_21h05m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-04_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-04_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-06_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-07_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-07_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-08_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-10_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-10_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-11_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-11_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-12_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-12_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-18_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-18_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-19_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-19_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-20_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-20_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-22_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-22_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-23_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-23_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-24_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-24_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-26_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-26_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-27_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-27_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-30_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-09-30_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-01_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-01_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-03_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-04_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-04_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-05_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-05_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-07_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-07_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-08_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-08_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-09_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-09_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-11_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-15_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-15_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-16_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-16_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-17_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-17_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-19_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-19_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-20_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-20_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-21_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-21_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-23_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-23_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-24_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-24_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-27_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-27_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-28_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-28_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-30_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-31_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-10-31_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-01_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-01_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-03_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-03_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-04_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-04_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-05_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-05_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-06_10h23m_336_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-07_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-07_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-08_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-08_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-09_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-09_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-11_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-12_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-12_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-13_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-15_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-15_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-16_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-16_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-17_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-17_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-19_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-19_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-20_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-20_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-23_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-23_21h05m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-24_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-24_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-26_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-27_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-27_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-28_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-28_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-30_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-11-30_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-01_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-01_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-02_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-02_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-04_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-05_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-05_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-06_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-06_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-08_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-08_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-09_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-09_20h50m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-10_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-10_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-12_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-12_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-13_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-13_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-14_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-16_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-16_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-17_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-17_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-20_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-20_21h05m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-21_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-21_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-23_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-24_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-24_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-25_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-25_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-27_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-27_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-28_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-28_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-29_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-29_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-31_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-12-31_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-01_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-01_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-02_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-02_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-04_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-04_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-05_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-05_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-06_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-06_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-08_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-08_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-09_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-09_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-10_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-10_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-12_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-12_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-13_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-13_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-16_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-16_21h05m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-17_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-17_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-19_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-20_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-20_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-21_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-21_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-23_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-23_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-24_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-24_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-25_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-25_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-27_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-27_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-28_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-28_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-29_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-29_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-31_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-01-31_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-01_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-02_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-02_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-04_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-05_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-05_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-06_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-06_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-08_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-08_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-12_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-12_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-13_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-13_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-15_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-16_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-16_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-17_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-17_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-19_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-19_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-20_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-20_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-21_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-21_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-23_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-23_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-24_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-24_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-25_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-25_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-27_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-27_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-28_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-02-28_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-01_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-01_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-03_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-03_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-04_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-04_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-05_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-05_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-07_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-07_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-08_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-08_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-11_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-11_21h05m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-12_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-14_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-15_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-15_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-16_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-16_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-18_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-18_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-19_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-19_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-20_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-20_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-22_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-22_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-23_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-23_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-24_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-24_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-26_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-26_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-27_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-27_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-28_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-28_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-30_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-30_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-31_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-03-31_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-01_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-01_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-03_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-03_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-04_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-04_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-07_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-07_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-08_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-08_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-10_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-11_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-11_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-12_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-12_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-14_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-14_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-15_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-15_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-16_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-16_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-18_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-18_21h21m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-19_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-19_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-20_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-20_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-04-22_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-04_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-04_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-05_09h17m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-05_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-07_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-08_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-08_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-09_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-09_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-11_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-11_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-12_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-12_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-13_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-13_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-15_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-15_21h21m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-16_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-16_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-17_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-17_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-19_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-19_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-20_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-24_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-24_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-25_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-25_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-27_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-27_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-28_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-28_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-31_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-05-31_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-01_09h17m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-01_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-03_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-04_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-04_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-05_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-05_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-07_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-07_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-08_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-08_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-09_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-09_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-11_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-11_21h21m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-12_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-12_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-13_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-13_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-15_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-15_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-16_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-16_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-17_09h02m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-17_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-19_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-19_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-20_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-20_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-21_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-21_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-23_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-24_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-24_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-27_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-27_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-28_09h17m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-28_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-06-30_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-01_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-01_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-02_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-02_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-04_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-04_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-05_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-05_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-06_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-06_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-08_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-08_21h21m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-09_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-09_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-10_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-10_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-12_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-12_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-13_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-13_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-14_09h02m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-14_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-16_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-16_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-17_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-17_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-18_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-18_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-20_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-20_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-21_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-21_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-22_20h18m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-24_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-24_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-25_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-25_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-27_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-28_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-28_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-29_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-29_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-07-31_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-01_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-01_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-02_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-04_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-05_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-05_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-06_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-06_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-08_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-08_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-09_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-09_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-10_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-10_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-12_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-12_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-13_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-13_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-14_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-14_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-16_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-16_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-17_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-17_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-19_10h09m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-19_21h32m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-20_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-20_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-21_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-21_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-23_10h05m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-23_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-24_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-24_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-25_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-25_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-27_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-27_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-28_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-28_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-29_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-29_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-31_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-08-31_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-01_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-01_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-02_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-02_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-04_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-04_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-05_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-05_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-06_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-06_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-08_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-08_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-09_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-09_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-10_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-10_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-12_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-12_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-15_10h09m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-15_21h32m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-17_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-25_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-25_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-26_20h09m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-27_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-27_21h21m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-28_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-28_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-29_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-29_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-09-30_20h05m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-01_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-01_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-02_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-02_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-03_09h02m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-03_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-04_10h16m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-04_20h01m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-05_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-05_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-06_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-06_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-07_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-07_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-08_10h13m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-08_19h58m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-09_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-09_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-10_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-10_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-11_08h54m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-12_10h09m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-12_21h32m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-13_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-13_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-14_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-14_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-16_10h05m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-16_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-17_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-17_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-18_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-18_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-19_20h12m_285_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-20_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-20_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-21_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-21_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-22_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-22_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-23_20h09m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-24_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-24_21h21m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-25_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-25_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-26_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-26_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-27_20h05m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-28_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-28_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-29_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-29_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-30_09h02m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-30_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-31_10h16m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-10-31_20h01m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-01_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-01_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-02_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-02_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-03_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-03_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-04_10h13m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-04_19h57m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-05_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-05_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-06_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-06_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-07_08h54m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-08_10h09m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-08_21h32m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-09_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-09_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-10_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-10_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-12_10h05m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-12_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-13_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-13_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-14_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-14_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-15_20h12m_285_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-16_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-16_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-17_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-17_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-18_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-18_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-19_20h09m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-20_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-20_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-21_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-21_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-22_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-22_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-23_20h05m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-24_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-24_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-25_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-25_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-26_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-26_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-27_10h16m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-27_20h01m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-28_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-28_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-29_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-29_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-30_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-11-30_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-01_10h13m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-01_19h57m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-02_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-02_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-03_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-03_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-04_08h54m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-05_10h09m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-05_21h32m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-06_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-06_21h05m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-07_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-07_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-09_10h05m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-09_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-10_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-10_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-11_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-11_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-12_20h12m_285_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-13_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-13_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-14_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-14_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-15_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-15_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-16_20h09m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-17_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-17_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-18_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-18_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-19_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-19_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-20_20h05m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-21_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-21_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-22_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-22_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-23_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-23_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-24_10h16m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-24_20h01m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-25_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-25_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-26_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-26_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-27_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-27_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-28_10h10m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-28_10h13m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-28_19h57m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-29_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-29_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-30_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-30_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2019-12-31_08h54m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-01_10h09m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-01_21h32m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-02_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-02_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-03_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-03_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-05_10h05m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-05_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-06_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-06_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-07_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-07_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-08_20h12m_285_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-09_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-09_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-10_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-10_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-11_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-11_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-12_20h09m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-13_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-13_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-14_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-14_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-15_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-15_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-16_20h05m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-17_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-17_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-18_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-18_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-19_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-19_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-20_10h16m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-20_20h01m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-21_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-21_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-22_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-22_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-23_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-23_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-24_10h13m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-24_19h57m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-25_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-25_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-26_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-26_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-27_08h54m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-28_10h09m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-28_21h32m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-29_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-29_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-30_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-01-30_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-01_10h05m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-01_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-02_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-02_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-03_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-07_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-08_20h09m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-09_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-09_21h21m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-10_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-10_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-11_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-11_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-12_20h05m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-13_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-13_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-14_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-14_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-15_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-15_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-16_10h16m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-16_20h01m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-17_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-17_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-18_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-18_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-19_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-19_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-20_10h13m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-20_19h57m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-21_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-21_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-22_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-22_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-23_08h54m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-24_10h09m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-24_21h32m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-25_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-25_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-26_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-26_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-28_10h05m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-28_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-29_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-02-29_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-01_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-01_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-02_20h12m_285_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-03_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-03_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-04_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-04_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-05_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-05_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-06_20h09m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-07_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-07_21h20m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-08_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-08_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-09_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-09_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-10_20h05m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-11_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-11_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-12_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-12_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-13_09h01m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-13_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-14_10h16m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-14_20h01m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-15_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-15_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-16_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-16_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-17_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-17_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-18_10h13m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-18_19h57m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-19_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-19_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-20_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-20_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-21_08h54m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-22_10h09m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-22_21h32m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-23_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-23_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-24_09h16m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-24_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-26_10h05m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-26_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-27_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-27_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-28_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-28_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-29_20h12m_285_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-30_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-30_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-31_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-03-31_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-01_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-01_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-02_20h09m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-03_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-03_21h21m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-04_09h31m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-04_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-05_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-05_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-06_20h05m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-07_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-07_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-08_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-08_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-09_09h02m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-09_20h24m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-10_10h16m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-10_20h01m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-11_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-11_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-12_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-12_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-13_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-13_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-14_10h13m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-14_19h58m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-15_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-15_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-16_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-16_20h43m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-17_08h54m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-18_10h09m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-18_21h32m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-19_09h43m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-19_21h06m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-20_09h17m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-20_20h39m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-22_10h05m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-22_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-23_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-23_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-24_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-24_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-25_20h13m_285_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-26_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-26_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-27_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-27_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-28_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-28_20h32m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-29_20h09m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-30_09h58m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-04-30_21h21m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-01_09h32m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-01_20h54m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-02_09h05m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-02_20h28m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-03_20h05m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-04_09h54m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-04_21h17m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-05_09h28m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-05_20h51m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-06_09h02m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-06_20h25m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-07_10h16m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-07_20h01m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-08_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-08_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-09_09h24m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-09_20h47m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-10_08h58m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-10_20h21m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-11_10h13m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-11_19h58m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-12_09h46m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-12_21h09m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-13_09h20m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-19_10h05m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-19_21h28m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-20_09h39m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-20_21h02m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-21_09h13m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-21_20h36m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-22_20h13m_285_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-23_10h01m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-23_21h24m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-24_09h35m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-24_20h58m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2020-05-25_09h09m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-02-27_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-01_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-01_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-02_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-02_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-03_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-03_20h34m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-05_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-05_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-06_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-06_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-07_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-07_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-09_09h56m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-09_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-10_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-10_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-11_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-11_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-13_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-13_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-14_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-14_20h49m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-15_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-15_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-17_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-18_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-18_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-19_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-21_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-21_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-22_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-22_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-25_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-26_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-26_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-28_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-28_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-29_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-29_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-30_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-03-30_20h34m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-01_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-01_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-02_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-02_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-03_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-03_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-05_09h56m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-05_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-06_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-06_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-07_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-07_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-09_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-09_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-10_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-10_20h49m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-11_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-15_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-17_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-17_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-18_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-18_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-20_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-21_09h41m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-21_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-04-22_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-04_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-04_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-06_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-06_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-07_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-07_20h49m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-08_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-08_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-10_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-10_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-11_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-11_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-12_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-14_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-14_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-15_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-15_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-17_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-18_09h41m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-18_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-19_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-19_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-21_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-21_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-22_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-22_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-23_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-23_20h34m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-25_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-25_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-26_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-26_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-27_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-27_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-29_09h55m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-29_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-30_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-30_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-31_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-05-31_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-02_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-02_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-03_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-03_20h48m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-04_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-04_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-06_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-06_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-07_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-07_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-08_20h18m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-10_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-10_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-11_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-11_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-13_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-14_09h40m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-14_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-15_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-15_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-17_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-17_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-18_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-18_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-19_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-19_20h33m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-21_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-21_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-22_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-22_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-23_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-25_09h55m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-25_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-26_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-26_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-27_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-27_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-29_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-29_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-30_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-06-30_20h48m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-01_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-01_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-03_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-03_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-04_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-04_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-05_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-07_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-07_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-08_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-08_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-10_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-11_09h40m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-11_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-12_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-12_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-14_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-14_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-15_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-15_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-16_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-16_20h33m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-18_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-18_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-19_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-19_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-20_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-20_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-22_09h55m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-22_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-23_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-23_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-24_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-24_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-26_09h49m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-26_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-26_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-27_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-27_20h48m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-28_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-28_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-30_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-30_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-07-31_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-01_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-03_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-03_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-04_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-04_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-06_10h07m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-06_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-07_09h41m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-07_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-08_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-08_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-10_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-10_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-11_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-11_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-12_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-12_20h33m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-14_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-14_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-15_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-15_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-16_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-16_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-18_09h55m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-18_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-19_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-19_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-20_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-20_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-22_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-22_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-23_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-23_20h48m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-24_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-24_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-26_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-26_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-27_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-27_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-28_08h56m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-28_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-30_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-30_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-31_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-08-31_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-02_10h07m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-02_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-03_09h40m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-03_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-04_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-04_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-06_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-06_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-07_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-07_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-08_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-08_20h33m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-10_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-10_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-11_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-11_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-12_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-12_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-15_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-15_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-17_20h03m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-18_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-18_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-25_10h10m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-25_19h55m_185_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-25_21h33m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-27_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-27_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-29_10h07m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-29_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-30_09h40m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-09-30_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-01_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-01_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-03_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-03_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-04_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-04_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-05_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-05_20h33m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-06_20h10m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-07_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-07_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-08_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-08_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-09_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-09_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-10_20h07m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-11_09h55m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-11_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-12_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-12_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-13_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-13_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-14_10h15m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-14_10h18m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-14_20h03m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-15_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-15_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-16_09h25m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-16_20h48m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-17_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-17_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-18_10h14m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-18_19h59m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-19_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-19_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-20_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-20_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-21_08h56m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-21_20h18m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-22_10h10m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-22_19h55m_185_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-22_21h33m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-23_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-23_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-24_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-24_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-26_10h07m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-26_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-27_09h40m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-27_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-28_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-28_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-30_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-30_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-31_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-10-31_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-01_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-01_20h33m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-02_20h10m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-03_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-03_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-04_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-04_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-05_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-05_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-06_20h07m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-07_09h52m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-07_09h55m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-07_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-08_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-08_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-09_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-09_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-10_10h18m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-10_20h03m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-11_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-11_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-12_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-12_20h48m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-13_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-13_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-14_10h14m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-14_19h59m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-15_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-15_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-16_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-16_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-17_08h56m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-17_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-18_10h10m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-18_19h55m_185_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-18_21h33m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-19_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-19_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-20_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-20_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-22_10h07m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-22_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-23_09h41m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-23_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-24_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-24_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-26_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-26_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-27_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-27_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-28_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-28_20h34m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-29_20h10m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-30_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-11-30_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-01_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-01_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-02_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-02_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-03_20h07m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-04_09h55m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-04_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-05_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-05_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-06_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-06_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-07_10h18m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-07_20h03m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-08_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-08_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-09_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-09_20h48m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-10_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-10_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-11_10h14m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-11_19h59m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-12_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-12_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-13_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-13_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-14_08h56m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-14_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-15_10h10m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-15_19h55m_185_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-15_21h33m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-16_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-16_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-17_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-17_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-19_10h07m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-19_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-20_09h41m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-20_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-21_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-21_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-23_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-23_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-24_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-24_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-25_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-25_20h33m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-26_20h07m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-26_20h10m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-27_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-27_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-28_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-28_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-29_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-29_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-30_20h07m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-31_09h55m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2019-12-31_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-01_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-01_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-02_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-02_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-03_10h18m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-03_20h03m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-04_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-04_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-05_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-05_20h48m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-06_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-06_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-07_10h14m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-07_19h59m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-08_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-08_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-09_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-09_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-10_08h56m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-10_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-11_10h10m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-11_19h55m_185_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-11_21h33m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-12_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-12_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-13_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-13_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-15_10h07m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-15_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-16_09h40m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-16_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-17_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-17_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-19_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-19_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-20_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-20_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-21_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-21_20h33m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-22_20h10m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-23_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-23_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-24_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-24_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-25_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-25_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-26_20h07m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-27_09h55m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-27_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-28_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-28_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-29_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-29_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-30_10h18m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-30_20h03m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-31_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-01-31_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-01_09h25m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-01_20h48m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-02_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-02_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-03_10h14m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-03_19h59m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-04_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-04_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-05_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-05_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-06_08h56m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-06_20h18m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-07_10h10m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-07_19h55m_185_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-07_21h33m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-08_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-08_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-09_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-09_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-11_10h07m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-11_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-12_09h40m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-12_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-13_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-13_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-15_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-15_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-16_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-16_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-17_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-17_20h33m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-18_20h10m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-19_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-19_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-20_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-20_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-21_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-21_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-22_20h07m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-23_09h55m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-23_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-24_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-24_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-25_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-25_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-26_10h18m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-26_20h03m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-27_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-27_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-28_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-28_20h48m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-29_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-02-29_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-01_10h14m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-01_19h59m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-02_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-02_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-03_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-03_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-04_08h56m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-04_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-05_10h10m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-05_19h55m_185_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-05_21h33m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-06_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-06_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-07_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-07_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-09_10h07m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-09_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-10_09h41m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-10_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-11_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-11_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-13_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-13_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-14_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-14_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-15_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-15_20h34m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-16_20h10m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-17_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-17_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-18_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-18_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-19_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-19_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-20_20h07m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-21_09h55m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-21_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-22_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-22_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-23_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-23_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-24_10h18m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-24_20h03m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-25_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-25_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-26_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-26_20h48m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-27_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-27_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-28_10h14m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-28_19h59m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-29_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-29_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-30_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-30_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-31_08h56m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-03-31_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-01_10h10m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-01_19h55m_185_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-01_21h33m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-02_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-02_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-03_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-03_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-05_10h07m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-05_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-06_09h41m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-06_21h03m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-07_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-07_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-09_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-09_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-10_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-10_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-11_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-11_20h34m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-12_20h10m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-13_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-13_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-14_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-14_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-15_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-15_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-16_20h07m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-17_09h56m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-17_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-18_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-18_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-19_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-19_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-20_10h18m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-20_20h03m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-21_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-21_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-22_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-22_20h49m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-23_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-23_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-24_10h14m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-24_19h59m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-25_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-25_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-26_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-26_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-27_08h56m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-27_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-28_10h11m_179_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-28_19h55m_185_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-28_21h33m_186_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-29_09h44m_193_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-29_21h07m_200_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-30_09h18m_207_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-04-30_20h41m_214_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-02_10h07m_236_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-02_21h30m_243_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-03_09h41m_250_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-03_21h04m_257_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-04_09h14m_264_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-04_20h37m_271_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-06_10h03m_293_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-06_21h26m_300_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-07_09h37m_307_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-07_21h00m_314_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-08_09h11m_321_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-08_20h34m_328_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-09_20h10m_342_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-10_09h59m_350_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-10_21h22m_357_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-11_09h33m_364_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-11_20h56m_371_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-12_09h07m_378_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-12_20h30m_385_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-13_20h07m_014_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-14_09h56m_022_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-14_21h18m_029_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-15_09h29m_036_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-15_20h52m_043_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-16_09h03m_050_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-16_20h26m_057_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-17_10h18m_065_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-17_20h03m_071_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-18_09h52m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-18_21h15m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-19_09h26m_093_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-19_20h49m_100_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-20_08h59m_107_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-20_20h22m_114_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-21_10h14m_122_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-21_19h59m_128_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-22_09h48m_136_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-22_21h11m_143_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-23_09h22m_150_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-23_20h45m_157_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-24_08h56m_164_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-24_20h19m_171_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3B_LST_2020-05-25_10h11m_179_DAY___1km_utm32_etrs89.tif']
+    files = file_list
+    #print(files)
+    #print(type(files))
+    #print(len(files))
+    #for i, files in enumerate(file_list):
+    for i in range(0, len(files)):
+        #print(files[i])
+        src1 = rio.open(files)
+        try:
+            for j, polygons in enumerate(import_list):
+                rio.mask.mask(src1, [import_list[0][j]], all_touched=0, crop=True, nodata=np.nan)
+            shutil.copy(files, tifs_selected)
+        except ValueError:
+            pass
+
+
+if __name__ == '__main__':
+    tifs_selected = directory + "/selected"
+    if os.path.exists(tifs_selected):
+        shutil.rmtree(tifs_selected)
+    os.mkdir(directory + "/selected")
+    file_list = extract_files_to_list(path_to_folder=directory)
+    #file_list = ['F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-03_09h50m_079_DAY___1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-03_21h13m_086_NIGHT_1km_utm32_etrs89.tif', 'F:/GEO411_data/Sentinel_Daten/S3A_LST_2018-07-04_20h47m_100_NIGHT_1km_utm32_etrs89.tif']
+    #print(file_list)
+
+    pool = mp.Pool(32)
+    print(pool.map(eliminate_nanoverlap, file_list))
