--- SenLAST/test_file.py
@@ -1,93 +1,0 @@
-from pathos import multiprocessing as mp
-from datetime import datetime
-import os
-import rasterio as rio
-import rasterio.mask
-import numpy as np
-import fiona
-import shutil
-
-start_time = datetime.now()
-
-### ----- TIFF Data ----- ###
-## Jonas Folder:
-# directory = "F:/Sentinel_3/S3_Daten-20200622T143538Z-001/S3_Daten"
-## Marlin Folder:
-directory = "F:/GEO411_data/Sentinel_Daten/"
-
-### ----- Shapefile Data ----- ###
-## Shapefile Marlin:
-shapefolder = "F:/GEO411_data/Daten_Sandra/new/"
-shapefile = "Stationen_Thüringen_Umland_3x3box.shp"
-## Shapefile Jonas:
-# shapefolder = "C:/Users/jz199/Documents/Studium/Master/2. Semester/Vorlesungsmitschriften/GEO411 - Landschaftsmanagement und Fernerkundung/Auszug_Daten_SandraBauer_MA/Auszug_Daten_SandraBauer_MA/"
-#shapefile = "Stationen_Thüringen_Umland_3x3box.shp"
-
-shape_path = shapefolder + shapefile
-
-
-def extract_files_to_list(path_to_folder):
-    print("a")
-    """
-    finds all .tif-files in the corresponding directory
-    :return:
-    """
-    new_list = []
-
-    for filename in os.listdir(path_to_folder):
-        if filename.endswith(".tif"):
-            new_list.append(os.path.join(path_to_folder, filename))
-        else:
-            continue
-    return new_list
-
-
-### this function needs to be generalized ###
-def import_polygons(shape_path):
-    print("b")
-    """
-    imports the 3x3km polygons of the DWD weather stations
-    :return:
-    """
-    shape_list = []
-    acitve_shapefile = fiona.open(shape_path, "r")
-    for i in range(0,len(list(acitve_shapefile))-1):
-        shapes = [feature["geometry"] for feature in acitve_shapefile]
-        shape_list.append(shapes)
-    #print(shape_list)
-    return shape_list
-
-
-def eliminate_nanoverlap(file_list):
-    print("c")
-    """
-    eliminates the scenes which would'nt match with all weather stations
-    :return:
-    """
-    #import_list = import_polygons(shape_path=shape_path)
-    import_list = [[{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}], [{'type': 'Polygon', 'coordinates': [[(657847.277237658, 5643474.213895654), (660847.277237658, 5643474.213895654), (660847.277237658, 5640474.213895654), (657847.277237658, 5640474.213895654), (657847.277237658, 5643474.213895654)]]}, {'type': 'Polygon', 'coordinates': [[(628925.3486363995, 5661625.574212378), (631925.3486363995, 5661625.574212378), (631925.3486363995, 5658625.574212378), (628925.3486363995, 5658625.574212378), (628925.3486363995, 5661625.574212378)]]}, {'type': 'Polygon', 'coordinates': [[(636133.5701770452, 5651253.4237408945), (639133.5701770452, 5651253.4237408945), (639133.5701770452, 5648253.4237408945), (636133.5701770452, 5648253.4237408945), (636133.5701770452, 5651253.4237408945)]]}, {'type': 'Polygon', 'coordinates': [[(602713.6356626084, 5635221.537168022), (605713.6356626084, 5635221.537168022), (605713.6356626084, 5632221.537168022), (602713.6356626084, 5632221.537168022), (602713.6356626084, 5635221.537168022)]]}, {'type': 'Polygon', 'coordinates': [[(685559.8310555458, 5593539.060318218), (688559.8310555458, 5593539.060318218), (688559.8310555458, 5590539.060318218), (685559.8310555458, 5590539.060318218), (685559.8310555458, 5593539.060318218)]]}, {'type': 'Polygon', 'coordinates': [[(631274.5275500177, 5623366.730672036), (634274.5275500177, 5623366.730672036), (634274.5275500177, 5620366.730672036), (631274.5275500177, 5620366.730672036), (631274.5275500177, 5623366.730672036)]]}, {'type': 'Polygon', 'coordinates': [[(596037.329981084, 5603435.850949816), (599037.329981084, 5603435.850949816), (599037.329981084, 5600435.850949816), (596037.329981084, 5600435.850949816), (596037.329981084, 5603435.850949816)]]}, {'type': 'Polygon', 'coordinates': [[(649866.626548125, 5597923.598440583), (652866.626548125, 5597923.598440583), (652866.626548125, 5594923.598440583), (649866.626548125, 5594923.598440583), (649866.626548125, 5597923.598440583)]]}, {'type': 'Polygon', 'coordinates': [[(623566.6801462888, 5614409.805070128), (626566.6801462888, 5614409.805070128), (626566.6801462888, 5611409.805070128), (623566.6801462888, 5611409.805070128), (623566.6801462888, 5614409.805070128)]]}, {'type': 'Polygon', 'coordinates': [[(653598.6058842163, 5614033.875900529), (656598.6058842163, 5614033.875900529), (656598.6058842163, 5611033.875900529), (653598.6058842163, 5611033.875900529), (653598.6058842163, 5614033.875900529)]]}, {'type': 'Polygon', 'coordinates': [[(607389.9233494792, 5640935.066472023), (610389.9233494792, 5640935.066472023), (610389.9233494792, 5637935.066472023), (607389.9233494792, 5637935.066472023), (607389.9233494792, 5640935.066472023)]]}, {'type': 'Polygon', 'coordinates': [[(663634.783650653, 5655931.27415407), (666634.783650653, 5655931.27415407), (666634.783650653, 5652931.27415407), (663634.783650653, 5652931.27415407), (663634.783650653, 5655931.27415407)]]}, {'type': 'Polygon', 'coordinates': [[(661602.2869801905, 5670669.269252692), (664602.2869801905, 5670669.269252692), (664602.2869801905, 5667669.269252692), (661602.2869801905, 5667669.269252692), (661602.2869801905, 5670669.269252692)]]}, {'type': 'Polygon', 'coordinates': [[(676169.2956554531, 5618969.54467234), (679169.2956554531, 5618969.54467234), (679169.2956554531, 5615969.54467234), (676169.2956554531, 5615969.54467234), (676169.2956554531, 5618969.54467234)]]}, {'type': 'Polygon', 'coordinates': [[(569324.5928862887, 5674510.354251895), (572324.5928862887, 5674510.354251895), (572324.5928862887, 5671510.354251895), (569324.5928862887, 5671510.354251895), (569324.5928862887, 5674510.354251895)]]}, {'type': 'Polygon', 'coordinates': [[(703268.0105129821, 5578799.93996204), (706268.0105129821, 5578799.93996204), (706268.0105129821, 5575799.93996204), (703268.0105129821, 5575799.93996204), (703268.0105129821, 5578799.93996204)]]}, {'type': 'Polygon', 'coordinates': [[(663951.8187899115, 5570760.306969032), (666951.8187899115, 5570760.306969032), (666951.8187899115, 5567760.306969032), (663951.8187899115, 5567760.306969032), (663951.8187899115, 5570760.306969032)]]}, {'type': 'Polygon', 'coordinates': [[(720546.8141810865, 5598393.641630656), (723546.8141810865, 5598393.641630656), (723546.8141810865, 5595393.641630656), (720546.8141810865, 5595393.641630656), (720546.8141810865, 5598393.641630656)]]}, {'type': 'Polygon', 'coordinates': [[(563433.6856855719, 5658483.407523126), (566433.6856855719, 5658483.407523126), (566433.6856855719, 5655483.407523126), (563433.6856855719, 5655483.407523126), (563433.6856855719, 5658483.407523126)]]}, {'type': 'Polygon', 'coordinates': [[(740969.6206497113, 5622231.113865188), (743969.6206497113, 5622231.113865188), (743969.6206497113, 5619231.113865188), (740969.6206497113, 5619231.113865188), (740969.6206497113, 5622231.113865188)]]}]]
-    # file_list = extract_files_to_list(path_to_folder=directory)
-    tifs_selected = directory + "/selected"
-    files = file_list
-    #for i, files in enumerate(file_list):
-    for i in range(0, len(files)):
-        print(files[i])
-        src1 = rio.open(files[i])
-        try:
-            for j, polygons in enumerate(import_list):
-                rio.mask.mask(src1, [import_list[0][j]], all_touched=0, crop=True, nodata=np.nan)
-            shutil.copy(files[i], tifs_selected)
-        except ValueError:
-            pass
-
-
-if __name__ == '__main__':
-    tifs_selected = directory + "/selected"
-    if os.path.exists(tifs_selected):
-        shutil.rmtree(tifs_selected)
-    os.mkdir(directory + "/selected")
-    file_list = extract_files_to_list(path_to_folder=directory)
-    print(file_list)
-
-    pool = mp.Pool(1)
-    print(pool.map(eliminate_nanoverlap, file_list))
